Further restructuring of core modules
Move getLength and canGetLength into core/length.js

Make the expecting object an attribute of wrap instead of a separately imported object
